requirements.txt
Flask
beautifulsoup4 
bs4
frozendict
geopy
ipython
kafka-python
matplotlib
numpy
py4j
pyelasticsearch
pymongo
requests
scipy
selenium
scikit-learn
tabulate
tldextract
wikipedia
findspark
iso8601
urllib3==1.24
joblib
flask-socketio==5.3.5
python-socketio==5.10.0
eventlet
